+++ test/plots/index.js
@@ -32,7 +32,6 @@ export {default as caltrainDirection} from "./caltrain-direction.js";
 export {default as carsDodge} from "./cars-dodge.js";
 export {default as carsHexbin} from "./cars-hexbin.js";
 export {default as carsJitter} from "./cars-jitter.js";
-export {default as carsLinearRegression} from "./cars-linear-regression.js";
 export {default as carsMpg} from "./cars-mpg.js";
 export {default as carsParcoords} from "./cars-parcoords.js";
 export {default as clamp} from "./clamp.js";
@@ -103,6 +102,8 @@ export {default as letterFrequencyColumn} from "./letter-frequency-column.js";
 export {default as letterFrequencyDot} from "./letter-frequency-dot.js";
 export {default as letterFrequencyLollipop} from "./letter-frequency-lollipop.js";
 export {default as letterFrequencyWheel} from "./letter-frequency-wheel.js";
+export {default as linearRegressionCars} from "./linear-regression-cars.js";
+export {default as linearRegressionPenguins} from "./linear-regression-penguins.js";
 export {default as likertSurvey} from "./likert-survey.js";
 export {default as logDegenerate} from "./log-degenerate.js";
 export {default as markovChain} from "./markov-chain.js";
@@ -141,7 +142,6 @@ export {default as penguinFacetDodgeIdentity} from "./penguin-facet-dodge-identi
 export {default as penguinFacetDodgeIsland} from "./penguin-facet-dodge-island.js";
 export {default as penguinFacetDodgeSymbol} from "./penguin-facet-dodge-symbol.js";
 export {default as penguinIslandUnknown} from "./penguin-island-unknown.js";
-export {default as penguinLinearRegression} from "./penguin-linear-regression.js";
 export {default as penguinMass} from "./penguin-mass.js";
 export {default as penguinMassSex} from "./penguin-mass-sex.js";
 export {default as penguinMassSexSpecies} from "./penguin-mass-sex-species.js";
